Move AppSkillButton component to matters dir
