+++ pages/index.vue
@@ -87,7 +87,7 @@ import AppHeader2 from '~/components/elements/AppHeader2'
 import AppLinkButton from '~/components/elements/AppLinkButton'
 import AppPayPayButton from '~/components/matters/AppPayPayButton'
 import AppProfileIcon from '~/components/AppProfileIcon'
-import AppSkillButton from '~/components/AppSkillButton'
+import AppSkillButton from '~/components/matters/AppSkillButton'
 import AppText from '~/components/elements/AppText'
 import AppWishListButton from '~/components/AppWishListButton'
 import AppWorks from '~/components/AppWorks'
